+++ components/mail/SMTP.php
@@ -63,7 +63,7 @@ class SMTP extends Base
      * @return \Zend\Mail\Message
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
-    public function getMessageClass($type)
+    protected function getMessageClass($type)
     {
         return new \Zend\Mail\Message();
     }
@@ -71,7 +71,7 @@ class SMTP extends Base
     /**
      * @return \Zend\Mail\Transport\TransportInterface
      */
-    public function getTransport()
+    protected function getTransport()
     {
         $options = [
             'name' => $this->name,